==> 02_BaMP_data_reconfig.R <==
# Following tidyr guide for data structure which is each variable (in this case
# protein LFQ intensity) as a column and each observation (sample) as a row. 
#
# Load tidyr which is what we'll be using further down to create batch and 
# anther length columns from the sample names.
library(tidyr)

# Load in the annotated table of protein LFQ intensities generated by the merge 
# function in 01_join_data.R
prot.data <- read.csv("output_tables/Proteomic_data.csv")
colnames(prot.data[,1:20])

# At the moment the data is in the opposite orientation we want (variables 
# as rows and observations as columns) and contains a lot of extra information.

# Subset the BaMP ID (protein) LFQ intensity and sample columns only.
bamp <- prot.data[,c(2,18:47)]

# If we flip the table now the row numbers will be the column headers:
# V1, V2, V3...
# To avoid this we'll make the BaMP (protein) IDs the rownames and remove the
# column.
rownames(bamp) <- bamp$BaMP
bamp <- bamp[,-1]
colnames(bamp[,1:5])

# Now we can invert the table using transpose "t()" in base R.
t.bamp <- t(bamp)
t.bamp <- as.data.frame(t.bamp)
colnames(t.bamp[,1:5])

# Remove "LFQ.Intensity." string from sample names (curently the rownames)
# and create a new column called "sample" to store the names using gsub
# (base R.
rownames(t.bamp)
t.bamp$sample <- gsub("LFQ.intensity.", # string to replace
                      "", # what to replace it with (nothing)
                      rownames(t.bamp)) # input string
t.bamp$sample

# Create a column for batch and anther length using tidyr's separate function.
t.bamp <- separate(t.bamp, # data frame
                   sample, # column to split
                   into = c("batch", "anther.len"), # names to give new columns
                   sep="_", # string that separates values to be split
                   remove = FALSE) # don't remove the column to split
t.bamp$batch
t.bamp$anther.len

# Re-order so that sample data comes before intensity data.
t.bamp <- t.bamp[,c(6432:6434, 1:6431)]
colnames(t.bamp[,1:5])

# Write it out to file.
write.csv(t.bamp, # data to write
          "output_tables/protein_vs_lfq.csv", # path write
          row.names = FALSE) # don't write out the row names
